Fix review; add debug mode for representations; add comments for blocks code; add separate attention.py
